[bench] Fix up rust benchmarks
Also add a CI step to build the benchmark code (without running them).
