+++ .github/workflows/rust.yml
@@ -52,6 +52,8 @@ jobs:
       - name: Run tests
         run: |
           cargo test --all-features
+      - name: Build benchmarks
+          cargo build --benches
   linux-arm:
     runs-on: buildjet-4vcpu-ubuntu-2204-arm
     timeout-minutes: 30
