+++ rust/benches/scan.rs
@@ -120,7 +120,7 @@ async fn create_file(path: &std::path::Path, mode: WriteMode) {
 
     let test_uri = path.to_str().unwrap();
     std::fs::remove_dir_all(test_uri).map_or_else(|_| println!("{} not exists", test_uri), |_| {});
-    let mut write_params = WriteParams {
+    let write_params = WriteParams {
         max_rows_per_file: num_rows as usize,
         max_rows_per_group: batch_size as usize,
         mode: mode,
